wemos-d1-r2/rotary_socketio/python_server/pyproject.toml
[tool.poetry]
name = "ws-rotary-server"
version = "0.1.0"
description = ""
authors = ["raushanraja <vraushanraja@gmail.com>"]
readme = "README.md"
packages = [{include = "ws_rotary_server"}]

[tool.poetry.dependencies]
python = "^3.10"
python-socketio = {extras = ["client"], version = "^5.8.0"}
requests = "^2.28.2"
eventlet = "^0.33.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
